llantos

+++ Node.cpp
@@ -1,6 +1,6 @@
 #include "Node.h"
 
-Node::Node(int x, int y, std::string AccentCity, int Population){
+Node::Node(float x, float y, std::string AccentCity, int Population){
     pos = Point(x, y);
     this->AccentCity = AccentCity;
     this->Population = Population;
