+++ Point.cpp
@@ -1,6 +1,6 @@
 #include "Point.h"
 
-Point::Point(int x, int y){
+Point::Point(float x, float y){
     this->x = x;
     this->y = y;    
 }
